+++ lib/mini_utils.dart
@@ -5,3 +5,5 @@ export 'src/dialogs/color_picker.dart' show showColorPicker;
 export 'src/widgets/login_screen.dart' show loginScreen;
 export 'src/widgets/digital_clock.dart' show digitalClock;
 export 'src/widgets/draggable_container.dart' show DraggableContainer;
+export 'src/widgets/line_connector_painter.dart'
+    show LineConnectorPainter, Tuple2;